feat(#1): Firebase 설정 및 로그인 기능 구현
- Firebase 프로젝트 생성 및 설정
- Authentication 활성화 (Email/Password, Google)
- Firestore Database 생성
- .env.local 파일 생성 (환경 변수 설정)
- use-auth Hook 개선 버전 적용
- 로그인/회원가입 페이지 개선 버전 적용
- Firestore CRUD 함수 null 체크 추가
- 타입 에러 수정



Closes #1
