+++ src/app/(auth)/signup/page-old.tsx
@@ -1,0 +1,167 @@
+'use client'
+
+import { useState } from 'react'
+import { useAuth } from '@/lib/hooks/use-auth'
+import { Button } from '@/components/ui/button'
+import { Input } from '@/components/ui/input'
+import { Label } from '@/components/ui/label'
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
+import { useToast } from '@/hooks/use-toast'
+import Link from 'next/link'
+
+export default function SignUpPage() {
+  const [email, setEmail] = useState('')
+  const [password, setPassword] = useState('')
+  const [confirmPassword, setConfirmPassword] = useState('')
+  const [isLoading, setIsLoading] = useState(false)
+  const { signUp, signInWithGoogle } = useAuth()
+  const { toast } = useToast()
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault()
+
+    if (password !== confirmPassword) {
+      toast({
+        variant: 'destructive',
+        title: '비밀번호 불일치',
+        description: '비밀번호가 일치하지 않습니다.',
+      })
+      return
+    }
+
+    if (password.length < 6) {
+      toast({
+        variant: 'destructive',
+        title: '비밀번호 오류',
+        description: '비밀번호는 최소 6자 이상이어야 합니다.',
+      })
+      return
+    }
+
+    setIsLoading(true)
+
+    try {
+      await signUp(email, password)
+      toast({
+        title: '회원가입 성공',
+        description: '대시보드로 이동합니다.',
+      })
+    } catch (error: any) {
+      toast({
+        variant: 'destructive',
+        title: '회원가입 실패',
+        description: error.message,
+      })
+    } finally {
+      setIsLoading(false)
+    }
+  }
+
+  const handleGoogleSignIn = async () => {
+    setIsLoading(true)
+    try {
+      await signInWithGoogle()
+      toast({
+        title: '회원가입 성공',
+        description: '대시보드로 이동합니다.',
+      })
+    } catch (error: any) {
+      toast({
+        variant: 'destructive',
+        title: 'Google 로그인 실패',
+        description: error.message,
+      })
+    } finally {
+      setIsLoading(false)
+    }
+  }
+
+  return (
+    <div className="flex min-h-screen items-center justify-center bg-gradient-to-br from-blue-50 to-indigo-100 p-4">
+      <Card className="w-full max-w-md">
+        <CardHeader className="space-y-1">
+          <CardTitle className="text-2xl font-bold text-center">회원가입</CardTitle>
+          <CardDescription className="text-center">
+            새 계정을 만들어 포트폴리오를 관리하세요
+          </CardDescription>
+        </CardHeader>
+        <CardContent>
+          <form onSubmit={handleSubmit} className="space-y-4">
+            <div className="space-y-2">
+              <Label htmlFor="email">이메일</Label>
+              <Input
+                id="email"
+                type="email"
+                placeholder="email@example.com"
+                value={email}
+                onChange={(e) => setEmail(e.target.value)}
+                required
+                disabled={isLoading}
+              />
+            </div>
+            <div className="space-y-2">
+              <Label htmlFor="password">비밀번호</Label>
+              <Input
+                id="password"
+                type="password"
+                placeholder="최소 6자 이상"
+                value={password}
+                onChange={(e) => setPassword(e.target.value)}
+                required
+                disabled={isLoading}
+              />
+            </div>
+            <div className="space-y-2">
+              <Label htmlFor="confirmPassword">비밀번호 확인</Label>
+              <Input
+                id="confirmPassword"
+                type="password"
+                placeholder="비밀번호를 다시 입력하세요"
+                value={confirmPassword}
+                onChange={(e) => setConfirmPassword(e.target.value)}
+                required
+                disabled={isLoading}
+              />
+            </div>
+            <Button
+              type="submit"
+              className="w-full"
+              disabled={isLoading}
+            >
+              {isLoading ? '가입 중...' : '회원가입'}
+            </Button>
+          </form>
+
+          <div className="relative my-4">
+            <div className="absolute inset-0 flex items-center">
+              <span className="w-full border-t" />
+            </div>
+            <div className="relative flex justify-center text-xs uppercase">
+              <span className="bg-background px-2 text-muted-foreground">
+                또는
+              </span>
+            </div>
+          </div>
+
+          <Button
+            variant="outline"
+            className="w-full"
+            onClick={handleGoogleSignIn}
+            disabled={isLoading}
+          >
+            Google로 가입하기
+          </Button>
+
+          <p className="mt-4 text-center text-sm text-muted-foreground">
+            이미 계정이 있으신가요?{' '}
+            <Link href="/login" className="text-primary underline-offset-4 hover:underline">
+              로그인
+            </Link>
+          </p>
+        </CardContent>
+      </Card>
+    </div>
+  )
+}
+
+
